bf16

+++ omni/speculative_train/scripts/train_qwen2_eagle.py
@@ -205,7 +205,7 @@ target_head = target_head.eval().npu().to(torch.bfloat16)
 print_with_rank("Initialized target head")
 
 config = AutoDraftModelConfig.from_file("/data/model/qwq-32b-eagle/config.json")
-draft_model = AutoEagleDraftModel.from_config(config).npu()
+draft_model = AutoEagleDraftModel.from_config(config).npu().to(torch.bfloat16)
 print(draft_model)
 names = [item[0] for item in draft_model.named_parameters()]
 print(names)
